# [semantic-arc] Teach the optimizer how to eliminate dead live ranges that are copied/consumed