… test/SILOptimizer/semantic-arc-opts.sil