The diagnose unreachable commit is just something that I hit as I was implementing this.

The main caveats here is that we do not handle load [take] that are forwarded through a forwarding instruction before the copy_value. I think with a bit of time, I can fix this.