@airspeedswift 2019-12-11
@swift-ci please benchmark