@gottesmm 2019-12-13
So I am going to use this PR for just a very simple version of this optimization that does not use a lot of compile time. Then I am going to do a more complex one.