@gottesmm 2019-12-15
I updated the PR with a much simpler version of this that can be expanded upon. Specifically:

1. I only optimize if the owned value has a single destroy, no forwarding uses, and no non-destroy uses that are reachable along any path that goes through the copy_value.
2. I only optimize where the consumed copy_value is in the same block as the dead live range of the original owned value.

The nice thing about this really simple version of the optimization is that I do not need to compute the LiveRange of our copy_value, just the @owned value. Assuming this fails, I can then compute the LiveRange of the copy_value and do more work.