@gottesmm 2019-12-15
@swift-ci test compiler performance