@gottesmm 2019-12-15
I am also going to do a force push with an additional test around properly viewing end_borrows as uses of the begin_borrow's operand along side the operand itself. Without that, we will miscompile.