@gottesmm 2019-12-16
Compiler perf bot didn't post the comment. This is the comment:

# Summary for master full

**Unexpected test results, excluded stats for RxCocoa, SwifterSwift, Base64CoderSwiftUI**

No regressions above thresholds

- [Debug-batch](#debug-batch-6212490da9bed506)
    - [debug-batch brief](#debug-batch-brief-6212490da9bed506)
    - [debug-batch detailed](#debug-batch-detailed-6212490da9bed506)
- [Release](#release-6212490da9bed506)
    - [release brief](#release-brief-6212490da9bed506)
    - [release detailed](#release-detailed-6212490da9bed506)



<a name="debug-batch-6212490da9bed506">

# Debug-batch

<a name="debug-batch-brief-6212490da9bed506">

## debug-batch brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (3)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 50,255,208,286,761 | 50,271,695,233,829 | 16,486,947,068 | 0.03%
LLVM.NumLLVMBytesOutput | 1,799,867,480 | 1,798,291,702 | -1,575,778 | -0.09%
time.swift-driver.wall | 5427.5s | 5408.9s | -18.6s | -0.34%
</details>

<a name="debug-batch-detailed-6212490da9bed506">

## debug-batch detailed
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (2)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Driver.NumDriverPipePolls | 79,695 | 78,354 | -1,341 | -1.68% :white_check_mark:
Driver.NumDriverPipeReads | 67,557 | 66,320 | -1,237 | -1.83% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (204)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ImportSetCacheHit | 1,796,211 | 1,796,032 | -179 | -0.01%
AST.ImportSetCacheMiss | 1,090,934 | 1,090,908 | -26 | -0.0%
AST.ImportSetFoldHit | 631,388 | 631,377 | -11 | -0.0%
AST.ImportSetFoldMiss | 459,546 | 459,530 | -16 | -0.0%
AST.ModuleShadowCacheHit | 1,548,378 | 1,548,483 | 105 | 0.01%
AST.ModuleShadowCacheMiss | 13,520 | 13,529 | 9 | 0.07%
AST.ModuleVisibilityCacheHit | 33,128 | 33,128 | 0 | 0.0%
AST.ModuleVisibilityCacheMiss | 8,399 | 8,399 | 0 | 0.0%
AST.NumASTBytesAllocated | 63,912,890,223 | 64,042,809,192 | 129,918,969 | 0.2%
AST.NumASTScopeLookups | 4,245,863 | 4,245,789 | -74 | -0.0%
AST.NumBraceStmtASTScopeExpansions | 624,966 | 624,966 | 0 | 0.0%
AST.NumBraceStmtASTScopes | 624,966 | 624,966 | 0 | 0.0%
AST.NumDecls | 139,809 | 139,809 | 0 | 0.0%
AST.NumDependencies | 420,736 | 420,738 | 2 | 0.0%
AST.NumInfixOperators | 53,667 | 53,667 | 0 | 0.0%
AST.NumIterableTypeBodyASTScopeExpansions | 269,858 | 269,859 | 1 | 0.0%
AST.NumIterableTypeBodyASTScopes | 352,063 | 352,048 | -15 | -0.0%
AST.NumLinkLibraries | 0 | 0 | 0 | 0.0%
AST.NumLoadedModules | 396,362 | 396,362 | 0 | 0.0%
AST.NumLocalTypeDecls | 250 | 250 | 0 | 0.0%
AST.NumLookupQualifiedInModule | 2,986,473 | 2,978,261 | -8,212 | -0.27%
AST.NumLookupQualifiedInNominal | 7,394,591 | 7,397,578 | 2,987 | 0.04%
AST.NumModuleLookupClassMember | 6,927 | 6,927 | 0 | 0.0%
AST.NumModuleLookupValue | 48,157,733 | 48,147,330 | -10,403 | -0.02%
AST.NumObjCMethods | 24,240 | 24,240 | 0 | 0.0%
AST.NumPostfixOperators | 49 | 49 | 0 | 0.0%
AST.NumPrecedenceGroups | 25,836 | 25,836 | 0 | 0.0%
AST.NumPrefixOperators | 99 | 99 | 0 | 0.0%
AST.NumReferencedDynamicNames | 189 | 189 | 0 | 0.0%
AST.NumReferencedMemberNames | 6,308,148 | 6,308,148 | 0 | 0.0%
AST.NumReferencedTopLevelNames | 476,000 | 476,000 | 0 | 0.0%
AST.NumSourceBuffers | 582,005 | 582,005 | 0 | 0.0%
AST.NumSourceLines | 4,716,580 | 4,716,580 | 0 | 0.0%
AST.NumSourceLinesPerSecond | 3,253,285 | 3,248,621 | -4,664 | -0.14%
AST.NumTotalClangImportedEntities | 6,205,996 | 6,207,868 | 1,872 | 0.03%
Driver.ChildrenMaxRSS | 243,399,575,552 | 242,656,413,696 | -743,161,856 | -0.31%
Driver.DriverDepCascadingDynamic | 0 | 0 | 0 | 0.0%
Driver.DriverDepCascadingExternal | 0 | 0 | 0 | 0.0%
Driver.DriverDepCascadingMember | 0 | 0 | 0 | 0.0%
Driver.DriverDepCascadingNominal | 0 | 0 | 0 | 0.0%
Driver.DriverDepCascadingTopLevel | 0 | 0 | 0 | 0.0%
Driver.DriverDepDynamic | 0 | 0 | 0 | 0.0%
Driver.DriverDepExternal | 0 | 0 | 0 | 0.0%
Driver.DriverDepMember | 0 | 0 | 0 | 0.0%
Driver.DriverDepNominal | 0 | 0 | 0 | 0.0%
Driver.DriverDepTopLevel | 0 | 0 | 0 | 0.0%
Driver.NumDriverJobsRun | 28,026 | 28,026 | 0 | 0.0%
Driver.NumDriverJobsSkipped | 0 | 0 | 0 | 0.0%
Driver.NumProcessFailures | 0 | 0 | 0 | 0.0%
Frontend.MaxMallocUsage | 1,215,853,855,080 | 1,215,904,611,072 | 50,755,992 | 0.0%
Frontend.NumInstructionsExecuted | 50,255,208,286,761 | 50,271,695,233,829 | 16,486,947,068 | 0.03%
Frontend.NumProcessFailures | 0 | 0 | 0 | 0.0%
IRModule.NumIRAliases | 196,255 | 196,257 | 2 | 0.0%
IRModule.NumIRBasicBlocks | 6,861,539 | 6,861,619 | 80 | 0.0%
IRModule.NumIRComdatSymbols | 0 | 0 | 0 | 0.0%
IRModule.NumIRFunctions | 3,361,071 | 3,360,774 | -297 | -0.01%
IRModule.NumIRGlobals | 3,671,597 | 3,671,577 | -20 | -0.0%
IRModule.NumIRIFuncs | 0 | 0 | 0 | 0.0%
IRModule.NumIRInsts | 87,955,308 | 87,872,763 | -82,545 | -0.09%
IRModule.NumIRNamedMetaData | 135,360 | 135,360 | 0 | 0.0%
IRModule.NumIRValueSymbols | 6,377,583 | 6,377,268 | -315 | -0.0%
LLVM.NumLLVMBytesOutput | 1,799,867,480 | 1,798,291,702 | -1,575,778 | -0.09%
Parse.NumFunctionsParsed | 268,908 | 268,908 | 0 | 0.0%
Parse.NumIterableDeclContextParsed | 922,412 | 922,438 | 26 | 0.0%
Parse.ParseAbstractFunctionBodyRequest | 246,318 | 246,318 | 0 | 0.0%
Parse.ParseMembersRequest | 756,485 | 756,511 | 26 | 0.0%
SILModule.NumSILGenDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenFunctions | 1,710,885 | 1,710,885 | 0 | 0.0%
SILModule.NumSILGenGlobalVariables | 53,546 | 53,546 | 0 | 0.0%
SILModule.NumSILGenVtables | 18,627 | 18,627 | 0 | 0.0%
SILModule.NumSILGenWitnessTables | 72,371 | 72,371 | 0 | 0.0%
SILModule.NumSILOptDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILOptFunctions | 2,462,444 | 2,462,442 | -2 | -0.0%
SILModule.NumSILOptGlobalVariables | 55,175 | 55,169 | -6 | -0.01%
SILModule.NumSILOptVtables | 31,424 | 31,424 | 0 | 0.0%
SILModule.NumSILOptWitnessTables | 158,535 | 158,535 | 0 | 0.0%
Sema.AbstractGenericSignatureRequest | 34,614 | 34,614 | 0 | 0.0%
Sema.AccessLevelRequest | 12,344,267 | 12,377,363 | 33,096 | 0.27%
Sema.AnyObjectLookupRequest | 281 | 281 | 0 | 0.0%
Sema.AreAllStoredPropertiesDefaultInitableRequest | 23,226 | 23,226 | 0 | 0.0%
Sema.AttachedFunctionBuilderRequest | 3 | 3 | 0 | 0.0%
Sema.AttachedPropertyWrapperTypeRequest | 539,524 | 539,524 | 0 | 0.0%
Sema.AttachedPropertyWrappersRequest | 2,209,167 | 2,209,164 | -3 | -0.0%
Sema.CallerSideDefaultArgExprRequest | 78,549 | 78,549 | 0 | 0.0%
Sema.ClassAncestryFlagsRequest | 109,163 | 109,163 | 0 | 0.0%
Sema.CollectOverriddenDeclsRequest | 7,377,271 | 7,379,105 | 1,834 | 0.02%
Sema.CompareDeclSpecializationRequest | 470,675 | 470,887 | 212 | 0.05%
Sema.CursorInfoRequest | 0 | 0 | 0 | 0.0%
Sema.CustomAttrNominalRequest | 3 | 3 | 0 | 0.0%
Sema.DefaultAndMaxAccessLevelRequest | 58,404 | 58,409 | 5 | 0.01%
Sema.DefaultArgumentExprRequest | 38,177 | 38,177 | 0 | 0.0%
Sema.DefaultArgumentInitContextRequest | 454 | 454 | 0 | 0.0%
Sema.DefaultDefinitionTypeRequest | 8,168 | 8,168 | 0 | 0.0%
Sema.DefaultTypeRequest | 455,892 | 455,892 | 0 | 0.0%
Sema.DynamicallyReplacedDeclRequest | 1,076,530 | 1,076,530 | 0 | 0.0%
Sema.EmittedMembersRequest | 27,969 | 27,969 | 0 | 0.0%
Sema.EnumRawTypeRequest | 25,550 | 25,550 | 0 | 0.0%
Sema.EnumRawValuesRequest | 10,581 | 10,581 | 0 | 0.0%
Sema.ExistentialConformsToSelfRequest | 14,336 | 14,365 | 29 | 0.2%
Sema.ExistentialTypeSupportedRequest | 17,717 | 17,717 | 0 | 0.0%
Sema.ExpandASTScopeRequest | 7,624,543 | 7,624,541 | -2 | -0.0%
Sema.ExtendedNominalRequest | 887,370 | 887,370 | 0 | 0.0%
Sema.ExtendedTypeRequest | 83,367 | 83,367 | 0 | 0.0%
Sema.FunctionBuilderTypeRequest | 3 | 3 | 0 | 0.0%
Sema.FunctionOperatorRequest | 72,937 | 72,937 | 0 | 0.0%
Sema.GenericParamListRequest | 12,764,463 | 12,783,721 | 19,258 | 0.15%
Sema.GenericSignatureRequest | 5,615,980 | 5,618,721 | 2,741 | 0.05%
Sema.GetDestructorRequest | 28,305 | 28,305 | 0 | 0.0%
Sema.HasCircularInheritanceRequest | 23,846 | 23,846 | 0 | 0.0%
Sema.HasCircularInheritedProtocolsRequest | 10,228 | 10,228 | 0 | 0.0%
Sema.HasCircularRawValueRequest | 8,958 | 8,958 | 0 | 0.0%
Sema.HasDefaultInitRequest | 56,819 | 56,819 | 0 | 0.0%
Sema.HasDynamicMemberLookupAttributeRequest | 707,990 | 708,165 | 175 | 0.02%
Sema.HasMemberwiseInitRequest | 20,343 | 20,343 | 0 | 0.0%
Sema.HasUserDefinedDesignatedInitRequest | 56,829 | 56,829 | 0 | 0.0%
Sema.InferredGenericSignatureRequest | 188,857 | 188,856 | -1 | -0.0%
Sema.InheritedDeclsReferencedRequest | 6,919,831 | 6,916,112 | -3,719 | -0.05%
Sema.InheritedTypeRequest | 313,411 | 313,486 | 75 | 0.02%
Sema.InheritsSuperclassInitializersRequest | 29,949 | 29,949 | 0 | 0.0%
Sema.InitKindRequest | 97,995 | 97,995 | 0 | 0.0%
Sema.InterfaceTypeRequest | 16,929,834 | 16,939,793 | 9,959 | 0.06%
Sema.IsABICompatibleOverrideRequest | 136,155 | 136,155 | 0 | 0.0%
Sema.IsAccessorTransparentRequest | 319,258 | 319,258 | 0 | 0.0%
Sema.IsCallableNominalTypeRequest | 2,089 | 2,089 | 0 | 0.0%
Sema.IsDeclApplicableRequest | 0 | 0 | 0 | 0.0%
Sema.IsDynamicRequest | 1,759,445 | 1,759,445 | 0 | 0.0%
Sema.IsFinalRequest | 2,673,472 | 2,675,908 | 2,436 | 0.09%
Sema.IsGetterMutatingRequest | 434,110 | 434,110 | 0 | 0.0%
Sema.IsImplicitlyUnwrappedOptionalRequest | 2,519,494 | 2,518,694 | -800 | -0.03%
Sema.IsObjCRequest | 1,594,266 | 1,594,473 | 207 | 0.01%
Sema.IsSetterMutatingRequest | 342,734 | 342,734 | 0 | 0.0%
Sema.IsStaticRequest | 3,748,147 | 3,749,963 | 1,816 | 0.05%
Sema.LazyStoragePropertyRequest | 2,550 | 2,550 | 0 | 0.0%
Sema.LookupInModuleRequest | 6,755,602 | 6,747,317 | -8,285 | -0.12%
Sema.LookupPrecedenceGroupRequest | 175,927 | 175,927 | 0 | 0.0%
Sema.MangleLocalTypeDeclRequest | 500 | 500 | 0 | 0.0%
Sema.NamedLazyMemberLoadFailureCount | 13,510 | 13,501 | -9 | -0.07%
Sema.NamedLazyMemberLoadSuccessCount | 31,636,828 | 31,628,618 | -8,210 | -0.03%
Sema.NamingPatternRequest | 205,987 | 205,984 | -3 | -0.0%
Sema.NeedsNewVTableEntryRequest | 694,555 | 694,555 | 0 | 0.0%
Sema.NominalTypeLookupDirectCount | 34,693,581 | 34,693,558 | -23 | -0.0%
Sema.NumAccessorBodiesSynthesized | 189,118 | 189,118 | 0 | 0.0%
Sema.NumAccessorsSynthesized | 294,770 | 294,770 | 0 | 0.0%
Sema.NumConformancesDeserialized | 9,566,668 | 9,576,901 | 10,233 | 0.11%
Sema.NumConstraintScopes | 27,686,496 | 27,688,639 | 2,143 | 0.01%
Sema.NumConstraintsConsideredForEdgeContraction | 92,802,517 | 92,802,295 | -222 | -0.0%
Sema.NumCyclicOneWayComponentsCollapsed | 0 | 0 | 0 | 0.0%
Sema.NumDeclsDeserialized | 76,416,753 | 76,447,611 | 30,858 | 0.04%
Sema.NumDeclsTypechecked | 1,411,738 | 1,411,738 | 0 | 0.0%
Sema.NumFunctionsTypechecked | 529,144 | 529,144 | 0 | 0.0%
Sema.NumGenericSignatureBuilders | 1,449,069 | 1,450,690 | 1,621 | 0.11%
Sema.NumLazyIterableDeclContexts | 9,716,637 | 9,718,704 | 2,067 | 0.02%
Sema.NumLazyRequirementSignatures | 1,027,126 | 1,027,322 | 196 | 0.02%
Sema.NumLazyRequirementSignaturesLoaded | 744,281 | 744,345 | 64 | 0.01%
Sema.NumLeafScopes | 17,648,805 | 17,650,674 | 1,869 | 0.01%
Sema.NumTypesDeserialized | 23,234,576 | 23,251,981 | 17,405 | 0.07%
Sema.NumTypesValidated | 1,521,046 | 1,521,042 | -4 | -0.0%
Sema.NumUnloadedLazyIterableDeclContexts | 6,237,555 | 6,233,753 | -3,802 | -0.06%
Sema.OpaqueReadOwnershipRequest | 281,382 | 281,382 | 0 | 0.0%
Sema.OpaqueResultTypeRequest | 0 | 0 | 0 | 0.0%
Sema.OperatorPrecedenceGroupRequest | 747 | 747 | 0 | 0.0%
Sema.OverriddenDeclsRequest | 3,041,838 | 3,044,083 | 2,245 | 0.07%
Sema.ParamSpecifierRequest | 1,305,216 | 1,305,215 | -1 | -0.0%
Sema.PatternBindingEntryRequest | 463,351 | 463,348 | -3 | -0.0%
Sema.PreCheckFunctionBuilderRequest | 0 | 0 | 0 | 0.0%
Sema.PropertyWrapperBackingPropertyInfoRequest | 534,155 | 534,155 | 0 | 0.0%
Sema.PropertyWrapperBackingPropertyTypeRequest | 539,524 | 539,524 | 0 | 0.0%
Sema.PropertyWrapperMutabilityRequest | 630,316 | 630,316 | 0 | 0.0%
Sema.PropertyWrapperTypeInfoRequest | 1 | 1 | 0 | 0.0%
Sema.ProtocolRequiresClassRequest | 75,876 | 75,999 | 123 | 0.16%
Sema.ProvideDefaultImplForRequest | 7,377,271 | 7,379,105 | 1,834 | 0.02%
Sema.RangeInfoRequest | 0 | 0 | 0 | 0.0%
Sema.RequirementRequest | 113,787 | 113,786 | -1 | -0.0%
Sema.RequirementSignatureRequest | 823,221 | 823,383 | 162 | 0.02%
Sema.RequiresOpaqueAccessorsRequest | 1,313,504 | 1,313,504 | 0 | 0.0%
Sema.RequiresOpaqueModifyCoroutineRequest | 272,298 | 272,298 | 0 | 0.0%
Sema.ResilienceExpansionRequest | 1,862,675 | 1,862,679 | 4 | 0.0%
Sema.ResolveImplicitMemberRequest | 372,805 | 373,200 | 395 | 0.11%
Sema.ResolveProtocolNameRequest | 0 | 0 | 0 | 0.0%
Sema.ResultTypeRequest | 693,025 | 693,028 | 3 | 0.0%
Sema.RootAndResultTypeOfKeypathDynamicMemberRequest | 0 | 0 | 0 | 0.0%
Sema.RootTypeOfKeypathDynamicMemberRequest | 0 | 0 | 0 | 0.0%
Sema.SelfAccessKindRequest | 3,150,666 | 3,153,778 | 3,112 | 0.1%
Sema.SelfBoundsFromWhereClauseRequest | 1,805,096 | 1,803,566 | -1,530 | -0.08%
Sema.SetterAccessLevelRequest | 145,155 | 145,155 | 0 | 0.0%
Sema.StorageImplInfoRequest | 1,420,497 | 1,420,497 | 0 | 0.0%
Sema.StoredPropertiesAndMissingMembersRequest | 32,387 | 32,387 | 0 | 0.0%
Sema.StoredPropertiesRequest | 353,778 | 353,778 | 0 | 0.0%
Sema.StructuralTypeRequest | 2,323 | 2,323 | 0 | 0.0%
Sema.SuperclassDeclRequest | 497,428 | 497,704 | 276 | 0.06%
Sema.SuperclassTypeRequest | 60,016 | 60,016 | 0 | 0.0%
Sema.SynthesizeAccessorRequest | 294,770 | 294,770 | 0 | 0.0%
Sema.SynthesizeDefaultInitRequest | 5,380 | 5,380 | 0 | 0.0%
Sema.SynthesizeMemberwiseInitRequest | 3,197 | 3,197 | 0 | 0.0%
Sema.TypeCheckFunctionBodyUntilRequest | 529,144 | 529,144 | 0 | 0.0%
Sema.TypeCheckSourceFileRequest | 27,072 | 27,072 | 0 | 0.0%
Sema.TypeDeclsFromWhereClauseRequest | 30,545 | 30,550 | 5 | 0.02%
Sema.TypeRelationCheckRequest | 0 | 0 | 0 | 0.0%
Sema.TypeWitnessRequest | 9,277 | 9,278 | 1 | 0.01%
Sema.USRGenerationRequest | 8,774,528 | 8,772,461 | -2,067 | -0.02%
Sema.UnderlyingTypeDeclsReferencedRequest | 382,533 | 382,218 | -315 | -0.08%
Sema.UnderlyingTypeRequest | 38,709 | 38,709 | 0 | 0.0%
Sema.UnqualifiedLookupRequest | 4,390,653 | 4,390,579 | -74 | -0.0%
Sema.ValueWitnessRequest | 68,497 | 68,497 | 0 | 0.0%
</details>

<a name="release-6212490da9bed506">

# Release

<a name="release-brief-6212490da9bed506">

## release brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (2)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 50,465,495,552,258 | 49,432,959,269,173 | -1,032,536,283,085 | -2.05% :white_check_mark:
time.swift-driver.wall | 9340.8s | 9225.9s | -114.9s | -1.23% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
LLVM.NumLLVMBytesOutput | 1,925,970,972 | 1,920,392,008 | -5,578,964 | -0.29%
</details>

<a name="release-detailed-6212490da9bed506">

## release detailed
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (18)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.NumLoadedModules | 30,053 | 30,053 | 0 | 0.0%
AST.NumTotalClangImportedEntities | 1,207,414 | 1,207,414 | 0 | 0.0%
IRModule.NumIRBasicBlocks | 5,448,505 | 5,450,799 | 2,294 | 0.04%
IRModule.NumIRFunctions | 2,823,716 | 2,822,597 | -1,119 | -0.04%
IRModule.NumIRGlobals | 3,182,311 | 3,181,417 | -894 | -0.03%
IRModule.NumIRInsts | 53,522,509 | 53,528,828 | 6,319 | 0.01%
IRModule.NumIRValueSymbols | 5,648,642 | 5,645,053 | -3,589 | -0.06%
LLVM.NumLLVMBytesOutput | 1,925,970,972 | 1,920,392,008 | -5,578,964 | -0.29%
SILModule.NumSILGenFunctions | 1,193,344 | 1,193,344 | 0 | 0.0%
SILModule.NumSILOptFunctions | 1,712,579 | 1,710,844 | -1,735 | -0.1%
Sema.NumConformancesDeserialized | 3,603,444 | 3,603,444 | 0 | 0.0%
Sema.NumConstraintScopes | 27,318,811 | 27,318,811 | 0 | 0.0%
Sema.NumDeclsDeserialized | 10,386,609 | 10,386,727 | 118 | 0.0%
Sema.NumFunctionsTypechecked | 532,766 | 532,766 | 0 | 0.0%
Sema.NumGenericSignatureBuilders | 280,435 | 280,435 | 0 | 0.0%
Sema.NumLazyIterableDeclContexts | 1,314,174 | 1,314,178 | 4 | 0.0%
Sema.NumTypesDeserialized | 5,266,312 | 5,266,274 | -38 | -0.0%
Sema.NumTypesValidated | 826,549 | 826,549 | 0 | 0.0%
</details>